+++ weblate/trans/tests/test_component.py
@@ -1058,3 +1058,54 @@ class ComponentEditMonoTest(ComponentEditTest):
 
         # It should be now read only
         self.assertEqual(source.unit_set.all()[0].state, STATE_READONLY)
+
+
+class ComponentKeyFilterTest(ViewTestCase):
+    """Test the key filtering implementation in Component."""
+
+    def create_component(self):
+        return self.create_android(key_filter="^tr")
+
+    def test_get_key_filter_re(self) -> None:
+        self.assertEqual(self.component.key_filter_re.pattern, "^tr")
+
+    def test_get_filtered_result(self) -> None:
+        translation = self.component.translation_set.get(language_code="en")
+        units = translation.unit_set.all()
+        self.assertEqual(units.count(), 1)
+        self.assertEqual(units.all()[0].context, "try")
+
+    def test_change_key_filter(self) -> None:
+        self.component.key_filter = "^th"
+        self.component.save()
+        self.assertEqual(self.component.key_filter_re.pattern, "^th")
+        translations = self.component.translation_set.all()
+        for translation in translations:
+            units = translation.unit_set.all()
+            self.assertEqual(units.count(), 1)
+            self.assertEqual(units.all()[0].context, "thanks")
+
+        self.component.key_filter = ""
+        self.component.save()
+        self.assertEqual(self.component.key_filter_re.pattern, "")
+        translations = self.component.translation_set.all()
+        for translation in translations:
+            units = translation.unit_set.all()
+            self.assertEqual(len(units), 4)
+
+    def test_bilingual_component(self):
+        project = self.component.project
+        component = self.create_po(
+            name="Bilingual Test", project=project, key_filter="^tr"
+        )
+        # Save should remove it
+        self.assertEqual(component.key_filter, "")
+        self.assertEqual(component.key_filter_re.pattern, "")
+
+        # Verify validation will reject it
+        component.key_filter = "^tr"
+        with self.assertRaisesMessage(
+            ValidationError,
+            "To use the key filter, the file format must be monolingual.",
+        ):
+            component.clean()